requirements.txt
aniso8601==8.0.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.18.0
pytz==2020.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.26.5
uuid==1.30
Werkzeug==1.0.1
zope.interface==5.4.0
